+++ RELEASE_NOTES.md
@@ -1,3 +1,13 @@
+### 1.1.93 - 13.10.2020
+* Added capability to load drivers from AppDomain, PR #694
+
+### 1.1.92 - 18.09.2020
+* MsSQL and MySQL: Removed computed column from Create parameters list (insert clause)
+
+### 1.1.91 - 04.06.2020
+* MySQL procedure calls fixed
+* Error handling and dll loading improvements
+
 ### 1.1.90 - 04.06.2020
 * DatabaseProviderTypes.MSSQLSERVER_DYNAMIC to support Microsoft.Data.SqlClient.dll
 * Improved SQLServer Udt parameters